libs/auth/pyproject.toml
[tool.poetry]
name = "adaptiv-auth"
version = "0.1.0"
description = "OIDC/JWT authentication utilities for Adaptiv-X services"
authors = ["Adaptiv-X Team"]
license = "Apache-2.0"
packages = [{include = "adaptiv_auth", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.0"
httpx = "^0.26.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
